+++ firmware/src/button.h
@@ -17,7 +17,7 @@ bool button_is_stuck();
 
 uint8_t button_num();
 void button_update();
-uint16_t button_read();
+uint32_t button_read();
 uint8_t button_real_gpio(int id);
 uint8_t button_default_gpio(int id);
